test: Better location for making queries to find invalid data
